extending corretly the tree of asserters
