+++ classes/mageekguy/atoum/stubs/asserters/asserter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace mageekguy\atoum\stubs\asserters;
+
+use mageekguy\atoum\stubs\asserters;
+
+abstract class asserter
+{
+	use asserters;
+}
