+++ classes/mageekguy/atoum/stubs/asserters/extension.php
@@ -9,6 +9,17 @@ class extension
     use asserters;
 
     /**
+     * Check if the extension is loaded (installed and enabled).
+     *
+     *    <?php
+     *    $this
+     *            ->extension('json')
+     *                    ->isLoaded()
+     *            ;
+     *
+     * Note: If you need to run tests only if an extension is present, you
+     *   can use the PHP annotation.
+     *
      * @param string $failMessage
      *
      * @return $this
