+++ classes/mageekguy/atoum/stubs/asserters/generator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace mageekguy\atoum\stubs\asserters;
+
+use mageekguy\atoum;
+
+class generator extends iterator
+{
+}
