+++ classes/mageekguy/atoum/stubs/asserters/phpResource.php
@@ -2,8 +2,6 @@
 
 namespace mageekguy\atoum\stubs\asserters;
 
-use mageekguy\atoum\stubs\asserters;
-
 class phpResource extends variable
 {
     /**
